ai_game_agent/requirements.txt
fastapi
uvicorn[standard]
anthropic
openai
requests
beautifulsoup4
pydantic
noise
Pillow
numpy
